requirements.txt
# HTTP
httpx>=0.27.0

# Config
pyyaml>=6.0
python-dotenv>=1.0.0

# Telegram Bot (async)
python-telegram-bot>=21.0

# Scheduler (async)
apscheduler>=3.10.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.28.0

# Market Calendar
exchange-calendars>=4.5.0

# Charts
matplotlib>=3.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
